Relocate util function to appropriate place
