--- packages/harpguru-core/src/components/toggle-buffer-flusher/utils/batch-toggle-degree-ids/index.ts
@@ -1,1 +1,0 @@
-export { batchToggleDegreeIds } from './batch-toggle-degree-ids'
