+++ packages/harpguru-core/src/utils/batch-toggle-degree-ids/batch-toggle-degree-ids.ts
