+++ packages/harpguru-core/src/utils/index.ts
@@ -1,3 +1,4 @@
+export { batchToggleDegreeIds } from './batch-toggle-degree-ids'
 export { getHarpFaceFacts } from './get-harp-face-facts'
 export { partiallyApplyNudgeFunction } from './paritally-apply-nudge-function'
 export { getScaledMenuLabelProtrusion } from './get-scaled-menu-label-protrusion'
